tests/test_app.py
# tests/test_app.py
import pytest
import os
from unittest.mock import patch, AsyncMock, MagicMock

# Adapt imports as needed
from app import create_openai_client, interact_with_openai, synthesize_and_play_speech, main

@pytest.mark.asyncio
async def test_interact_with_openai_error_handling():
    client = AsyncMock()
    prompts = [{"role": "user", "content": None}]  # Invalid prompt
    response = await interact_with_openai(client, prompts)
    print(response)
    assert "Error in the AI response" not in response

pytest.mark.asyncio
async def test_main_flow():
    # Mock the main function dependencies
    # Assume main executes all major components at least once
    with patch('app.create_openai_client') as mock_create_client:
        client_instance = AsyncMock()
        mock_create_client.return_value = client_instance
        client_instance.chat.completions.create.return_value = AsyncMock(choices=[AsyncMock(message=AsyncMock(content="Hello"))])
        with patch('app.transcribe_speech_to_text', return_value="This is a test"):
            with patch('app.synthesize_and_play_speech') as mock_synth:
                await main()
                mock_synth.assert_called_once()

@pytest.mark.asyncio
async def test_invalid_client_creation():
    # Set up the environment for testing missing API key
    os.environ['AZURE_OPENAI_API_KEY'] = ''
    # The other variables can be set to dummy values
    os.environ['AZURE_API_VERSION'] = 'dummy-version'
    os.environ['AZURE_OPENAI_ENDPOINT'] = 'dummy-endpoint'

    with pytest.raises(ValueError) as exc_info:
        await create_openai_client()

    assert "AZURE_OPENAI_API_KEY is required but missing" in str(exc_info.value), "Should raise ValueError for missing API key"
